requirements.txt
streamlit==1.18.1
langchain==0.0.154
python-dotenv==1.0.0
PyPDF2==3.0.1
streamlit-extras==0.2.7
openai==0.27.8
tiktoken==0.4.0
faiss-cpu==1.7.4
altair==4.0.0